refactor(278): inline import routes observer logic
